+++ KrkrDataLoader/src/main/java/KrkrDataLoader/Main.java
@@ -22,9 +22,9 @@ public class Main
 		// TODO：准备开始开发前端内容
 		GlobalSetting.loadFromJson();
 		
-//		String scenePath = "KrkrDataLoader/src/test/resources/001・アーサー王ver1.07.ks.json";
-//		ConfigCommander configCommander = new ConfigCommander(scenePath);
-//		configCommander.commandMain();
+		String scenePath = "KrkrDataLoader/src/test/resources/001・アーサー王ver1.07.ks.json";
+		ConfigCommander configCommander = new ConfigCommander(scenePath);
+		configCommander.commandMain(System.in,System.out);
 
 //		KrkrScenes scenes = new KrkrScenes(scenePath);
 //		System.out.println(scenes.getName());
@@ -46,9 +46,9 @@ public class Main
 //		}
 		
 
-		System.setProperty("server.port", "8080");
-
-		SpringApplication.run(Main.class, args);
+//		System.setProperty("server.port", "8080");
+//
+//		SpringApplication.run(Main.class, args);
 	}
 	
 	@Bean